cleaned up offboarding cron
