+++ src/crons/off-boarding/index.ts
@@ -1,13 +1,12 @@
 import * as dotenv from 'dotenv-safe';
 dotenv.config();
 
-import { isToday } from '../../utils/dates';
-import { getUsersAtRc } from '../../recurse-api';
 import {
-  offBoardUsers,
-  notifyAdminOffboardingResults,
+  getUsersToOffBoard,
+  deactivateUsers,
   notifyDeactivatedUsers
-} from '../../crons/off-boarding/offboard-users';
+} from './offboardinghelpers';
+import { notifyAdmin } from '../../zulip-messenger';
 
 export async function handlePossibleOffboarding() {
   const users = await getUsersToOffBoard();
@@ -18,21 +17,26 @@ export async function handlePossibleOffboarding() {
 
   console.log('offboarding ', users.length, 'users on', new Date());
 
-  const { error, success } = offBoardUsers(users.map(user => user.email));
-
-  notifyAdminOffboardingResults(error, success);
-  notifyDeactivatedUsers(success);
-}
-
-export const getUsersToOffBoard = async () =>
-  getUsersAtRc().then(users =>
-    users.filter(user =>
-      // NOTE: faculty members have stints with end_date = null.
-      user.stints.some(
-        stint =>
-          !!stint.end_date &&
-          stint.type === 'retreat' &&
-          isToday(stint.end_date)
-      )
-    )
+  const { deactivatedUsers, errorUsers } = deactivateUsers(
+    users.map(user => user.email)
   );
+
+  notifyDeactivatedUsers(deactivatedUsers);
+
+  // Notify Admin users who were offboarded and users who could not be offboarded
+  notifyAdmin(
+    `${
+      process.env.NODE_ENV === 'production'
+        ? 'PROD: offboarded'
+        : 'DEV: Would offboard'
+    } the following users: ${JSON.stringify(deactivatedUsers)}`,
+    'LOG'
+  );
+
+  if (errorUsers.length) {
+    notifyAdmin(
+      `ERROR off boarding the following users: ${JSON.stringify(errorUsers)}`,
+      'WARNING'
+    );
+  }
+}
